enable to customize SQL comment prefix
In our use case, it would be very handy to be able to have the SQL
comments formatted as correlation_id=XXX correlation_id=YYY (we may have
more than one cid for the same request).
We are using Splunk and this would provide auto-indexing for the
correlation_id field.