setup.py: Fix long_description
PyPI does not allow the `raw` directive in the long description. I
want to keep the "backward-incompatible" tag in the HTML
documentation, it really stands out. So let's tweak the contents of
the HISTORY.txt file instead.

If you see this commit message in a few months when it breaks, here is
my message: I did not really think that it would not break. Sorry!