middleware: Don't reuse correlation id across requests processed by the same thread
This is a partial revert of 4476392b4e6c0d9f9a38074359cb139f6612f77b.

That commit introduced a bug that caused the correlation id to persist
across requests if the same thread was used to process them (which is
the case with uWSGI for example). All these requests had the same
correlation id.

The intention of the reverted commit is lost: we cannot use `get_cid`
anymore when the middleware is not used.

A better fix would probably be to get rid of the `_process_request()`
part of the middleware. Instead, we could register a signal handler
for "request_started" that would set the correlation id. That way, we
could greatly simplify the middleware and make it only set the
response header if configured.