cid: Don't reuse correlation id across requests processed by the same thread
This is a partial revert of 4476392.

That commit introduced a bug that caused the correlation id to persist
across requests if the same thread was used to process them (which is
the case with uWSGI for example). All these requests had the same
correlation id.

The intention of the reverted commit is not lost. If you don't use the
middleware, you should now call `cid.locals.generate_new_cid` before
trying to call `get_cid`.