Stop including changelog in PyPI long description
We were already trying hard to tweak the contents of the changelog to
make it somewhat readable in PyPI long description. Now that we are
declaring a "strike" role in the changelog, it's becoming too hard.
And the changelog of versions 2.2 and 2.3 would be very confusing
without the striked parts.

A future commit will add a link to the changelog from the PyPI page.
That should be enough.